requirements-devel.txt
nose
lettuce
pyhamcrest
